% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_build.R
\name{excel_build}
\alias{excel_build}
\title{Generate a formated Excel File from large dataframe}
\usage{
excel_build(model = NULL)
}
\arguments{
\item{model}{a model object generated by}
}
\description{
This is an internal function used by process() to subset and output a large dataframe as a formated multisheet workbook named after the rainfall distirbution and depth i.e. 'traingular_0.4m.xlsx'.
This function heavily utilizes the r2excel functions developed by Alboukadel Kassambara (https://github.com/kassambara).
Sheet 1 of each output workbook is a README documnet detailing the date, time, and user who generated the workbook; the modeled event and hill slope; and the contents of the remaining 8 sheets.
}
\examples{
excel_bulid(triangular_rainfall_mod)
}
\author{
Mike Johnson
}
