% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe.R
\name{exe}
\alias{exe}
\title{Execute Model}
\usage{
exe(scenario_file, save.as = FALSE)
}
\arguments{
\item{save.as}{logical. If TRUE model data will be saved as an RDS file in the working directory. Default is FALSE}

\item{sceanrio_file}{A three element list built using param()}
}
\description{
Run the model for a sceanrio file built using param().
}
\examples{
mod = exe(scene1, save.as = TRUE)
}
\author{
Mike Johnson
}
