% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rainfall.R
\name{generate_rainfall}
\alias{generate_rainfall}
\title{Generate a rainfall distirbution}
\usage{
generate_rainfall(type = "input", depth = NULL, hours = NULL,
  timestep = NULL, path = NULL)
}
\arguments{
\item{type}{character. Used to define the distribution of rainfall. Options are: (1) Constant (2) triangualar (3) input}

\item{depth}{numeric. Describes the total depth of rainfall in meters}

\item{hours}{numeric. Decribes the total duration of rainfall in hours}

\item{timestep}{numeric. Describes the timestep to generate rainfall measurements in seconds}

\item{path}{character. Describes a path to an input csv or xlsx file if type is 'input'}
}
\description{
This is an internal function used by param() to generate a rainfall timeseries based on input values.
}
\examples{
constant_ppt = generate_rainfall(type = 'constant', depth = .4, hours = 24, timestep = 1, path = NULL)
triangular_ppt = generate_rainfall(type = 'triangular', depth = .4, hours = 24, timestep = 1, path = NULL)
nrcs_ppt = generate_rainfall(type = 'input', depth = NULL, hours = NULL, timestep = NULL, path = "nrcs_type2_0.40m.xlsx")

}
\author{
Mike Johnson
}
