% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param}
\alias{param}
\title{Parameterize a Scenario file}
\usage{
param()
}
\value{
This function will return a three element list (hill, soil, rain) accesable via the '$' operator
}
\description{
This function paramaterizes a scenario file for the model via a GUI. This is necessary to ensure the file is formated correctly for a model run. To use, simply declare a varaible and run param().
}
\details{
Upon execution, a GUI interface will appear and request the following infomation to define three lists: (1) Hill (2) Soil (3) rainfall
If you have already declared a sceanrio file, simply type the variable name in the approapitate heading and click 'Use old run'

Hill shape

Vertical Units: Provide in units of (m) (i.e. 60)

Horizontal Units: Provide in units of (m) (i.e. 300)

Width Units: Provide in units of (m) (i.e. 50)

Roughness: Provide in units of () (i.e. )

Conductivity: Provide in units of () (i.e. )

Soil Water Tension: Provide in units of () (i.e. )

Porosity: Provide in units of () (i.e. )

Water Content: Provide in units of () (i.e. )

A user is given three options for defining a rainfall distribution. (1) A constant rate (2) a triangular distribution (3) an input file\n

In all cases the user must specify a type as: 'Constant', 'Triangular', or 'Input'
For cases 1 and 2 a user must also specify the rainfall

Duration: Provide in units of (hours) (i.e. 24 )

Depth: Provide in units of (m) (i.e. .4 )

Timestep: Provide in units of (seconds) (i.e. 1)

For an input storm, a user must specify a path to a .csv or xlsx file

Finally, specify a distance (deltaX) and time step (deltaT) for the model to calculate over.

Once complete submit the information to store and derive all needed model parameters to the decalred variable. If you have already declared a sceanrio file, simply type the variable name in the approapitate heading and select 'Use old run' to auto populate the scenario file.
}
\examples{
scene1 = param()
}
\author{
Mike Johnson
}
