% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Process Model Object}
\usage{
process(model = NULL, hr = NULL, min = NULL, sec = NULL, plot = TRUE,
  excel = FALSE)
}
\arguments{
\item{model}{a model file calcualted with exe()}

\item{hr}{numeric. Define timesteps of interest in hours. Leave min and sec as NULL}

\item{min}{numeric. Define timesteps of interest in min. Leave hr and sec as NULL}

\item{sec}{numeric. Define timesteps of interest in seconds. Leave hr and min as NULL}

\item{plot}{logical. Request plots be generated for each varaible at each timestep. If TRUE plots will be stored as an element 'plot' in the output list object. Default is FALSE.}

\item{excel}{logical. Request a formated excel workbook be generated for processed data. Returned workbook contains a README sheet detailing the date, time, and user who generated the workbook; the modeled event and hill slope; and the contents of the remaining 8 sheets.}
}
\value{
A list with elements for:

1) Scenario file

2) Runoff depth (y)

3) Infiltration (F)

4) Wetting Front (zf)

5) Runoff Volume (q)

6) Runoff depth at toe (y_toe)

7) Plots (if Plots = TRUE)
}
\description{
Used to process and subset a model run generate with exe(). Here you can define the timesteps of interest and request a dataframe, excel file or graphical output t for each variable at those timesteps.
Timesteps can be defined by a vector of hours, minutes or seconds.
}
\examples{
constant_sub = process(model = constant_mod, hr = c(0,5,10,15,20,24), min = NULL, sec = NULL, plot = TRUE, excel = FALSE)
triangular_sub = process(model = triangular_mod, hr = NULL, min = NULL, sec = c(1200, 6000, 25200, 52200, 68100), plot = TRUE, excel = TRUE)

}
\author{
Mike Johnson
}
