% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_runoff.R
\name{time_of_runoff}
\alias{time_of_runoff}
\title{Determine time of inital runoff}
\usage{
time_of_runoff(scenario_file)
}
\arguments{
\item{scenario_file}{a scenario_file generated with param()}
}
\description{
This is an internal function used by param() to calcualte time of initial runoff from a rainfall distribution generated with generate_rainfall().
}
\examples{
tp = time_of_runoff(scene1)

}
\author{
Mike Johnson
}
